Decouple run and diagnostics from adapters
Move SensorFrame into shared/types and move JSONL run-log decoding into shared/boundaries so use_cases no longer import adapter-owned transport or persistence modules. Refactor sample_builder to consume SpeedContext inputs, update tests/docs, and remove the resolved layer-boundary allowlist entries.
