+++ apps/server/tests/hygiene/test_layer_boundaries.py
@@ -87,7 +87,6 @@ _KNOWN_VIOLATIONS: frozenset[tuple[str, str]] = frozenset(
         ("domain/strength_metrics.py", "vibesensor.coerce"),
         # shared → adapters
         # use_cases → adapters
-        ("use_cases/diagnostics/helpers.py", "vibesensor.adapters.persistence.runlog"),
         ("use_cases/history/exports.py", "vibesensor.adapters.persistence.history_db"),
         ("use_cases/history/helpers.py", "vibesensor.adapters.persistence.history_db"),
         ("use_cases/history/reports.py", "vibesensor.adapters.persistence.history_db"),
@@ -95,8 +94,6 @@ _KNOWN_VIOLATIONS: frozenset[tuple[str, str]] = frozenset(
         ("use_cases/run/logger.py", "vibesensor.adapters.gps.gps_speed"),
         ("use_cases/run/logger.py", "vibesensor.adapters.persistence.history_db"),
         ("use_cases/run/post_analysis.py", "vibesensor.adapters.persistence.history_db"),
-        ("use_cases/run/sample_builder.py", "vibesensor.adapters.udp.protocol"),
-        ("use_cases/run/sample_builder.py", "vibesensor.adapters.gps.gps_speed"),
         # use_cases → infra
         ("use_cases/history/reports.py", "vibesensor.infra.config.settings_store"),
         ("use_cases/history/runs.py", "vibesensor.infra.config.settings_store"),
