+++ apps/server/tests/use_cases/diagnostics/test_coverage_gap_audit_top10.py
@@ -681,12 +681,19 @@ class TestExtractStrengthData:
 class TestResolveSpeedContext:
     """Tests for _resolve_speed_context via a minimal RunRecorder setup."""
 
+    @staticmethod
+    def _resolve_from_logger(logger) -> tuple[float | None, float | None, str, float | None]:
+        resolution = logger.gps_monitor.resolve_speed()
+        return resolve_speed_context(
+            gps_speed_mps=logger.gps_monitor.speed_mps,
+            resolved_speed_mps=resolution.speed_mps,
+            resolved_speed_source=resolution.source,
+            analysis_settings_snapshot=logger._analysis_settings_snapshot(),
+        )
+
     def test_no_speed_available(self) -> None:
         logger, _ = _make_run_recorder()
-        speed_kmh, gps_speed, source, rpm = resolve_speed_context(
-            logger.gps_monitor,
-            logger._analysis_settings_snapshot(),
-        )
+        speed_kmh, gps_speed, source, rpm = self._resolve_from_logger(logger)
         assert speed_kmh is None
         assert rpm is None
 
@@ -694,10 +701,7 @@ class TestResolveSpeedContext:
         logger, gps_mock = _make_run_recorder()
         gps_mock.speed_mps = 10.0  # 36 km/h
         gps_mock.resolve_speed.return_value = MagicMock(source="gps", speed_mps=10.0)
-        speed_kmh, gps_speed, source, rpm = resolve_speed_context(
-            logger.gps_monitor,
-            logger._analysis_settings_snapshot(),
-        )
+        speed_kmh, gps_speed, source, rpm = self._resolve_from_logger(logger)
         assert speed_kmh == pytest.approx(36.0, rel=0.01)
         assert gps_speed == pytest.approx(36.0, rel=0.01)
         assert source == "gps"
@@ -707,10 +711,7 @@ class TestResolveSpeedContext:
         logger, gps_mock = _make_run_recorder()
         gps_mock.override_speed_mps = 20.0  # 72 km/h
         gps_mock.resolve_speed.return_value = MagicMock(source="manual", speed_mps=20.0)
-        speed_kmh, _, source, _ = resolve_speed_context(
-            logger.gps_monitor,
-            logger._analysis_settings_snapshot(),
-        )
+        speed_kmh, _, source, _ = self._resolve_from_logger(logger)
         assert speed_kmh == pytest.approx(72.0, rel=0.01)
         assert source == "manual"
 
@@ -725,10 +726,7 @@ class TestResolveSpeedContext:
             rim_in=16,
             final_drive_ratio=3.73,
         )
-        _, _, _, rpm = resolve_speed_context(
-            logger.gps_monitor,
-            logger._analysis_settings_snapshot(),
-        )
+        _, _, _, rpm = self._resolve_from_logger(logger)
         assert rpm is None, "Without gear_ratio, RPM should not be estimated"
 
     def test_uses_order_reference_spec_for_engine_rpm(
@@ -748,10 +746,7 @@ class TestResolveSpeedContext:
         gps_mock.speed_mps = 10.0
         gps_mock.resolve_speed.return_value = MagicMock(source="gps", speed_mps=10.0)
 
-        speed, _, _, rpm = resolve_speed_context(
-            logger.gps_monitor,
-            logger._analysis_settings_snapshot(),
-        )
+        speed, _, _, rpm = self._resolve_from_logger(logger)
 
         assert speed == pytest.approx(36.0, rel=0.01)
         assert rpm == 1234.5
